Isolate string functions into their own runtime module
