+++ opal/runtime/boot.js
@@ -2175,78 +2175,6 @@
     return Opal.load_normalized(path);
   };
 
-
-  // Strings
-  // -------
-
-  // We use a helper to create new Strings, globally, so that it
-  // will be easer to change that later on to a mutable string class.
-  // Also this helper always sets a encoding. If encoding is not
-  // provided, "UTF-8" will be used.
-  // @returns a new String object with encoding set.
-  function $str(str, encoding) {
-    if (!encoding || encoding === nil) encoding = "UTF-8";
-    str = Opal.set_encoding(new String(str), encoding);
-    str.internal_encoding = str.encoding;
-    return str;
-  }
-  Opal.str = $str;
-
-  // Provide the encoding register with a default "UTF-8" encoding, because
-  // its used from the start and will be raplaced by the real UTF-8 encoding
-  // when 'corelib/string/encoding' is loaded.
-  Opal.encodings = { __proto__: null, "UTF-8": { name: "UTF-8" }};
-
-  // Sets the encoding on a string, will treat string literals as frozen strings
-  // raising a FrozenError.
-  //
-  // @param str [String] the string on which the encoding should be set
-  // @param name [String] the canonical name of the encoding
-  // @param type [String] possible values are either `"encoding"`, `"internal_encoding"`, or `undefined
-  Opal.set_encoding = function(str, name, type) {
-    if (typeof type === "undefined") type = "encoding";
-    if (typeof str === 'string' || str.$$frozen === true)
-      $raise(Opal.FrozenError, "can't modify frozen String");
-
-    let encoding = Opal.find_encoding(name);
-    if (encoding === str[type]) { return str; }
-    str[type] = encoding;
-
-    return str;
-  };
-
-  // Fetches the encoding for the given name or raises ArgumentError.
-  Opal.find_encoding = function(name) {
-    let register = Opal.encodings;
-    let encoding = register[name] || register[name.toUpperCase()];
-    if (!encoding) $raise(Opal.ArgumentError, "unknown encoding name - " + name);
-    return encoding;
-  }
-
-  Opal.fallback_to_s = function(obj) {
-    return `#<${obj.$$class.$to_s()}:0x${Opal.id(obj).toString(16)}>`
-  }
-
-  Opal.to_s = function(obj) {
-    var stringified;
-    if (obj == null) {
-      return "`"+String(obj)+"`";
-    }
-    else if (typeof obj === 'string' || (typeof obj === 'object' && obj.$$is_string)) {
-      return obj;
-    }
-    else if (obj.$to_s != null && !obj.$to_s.$$stub) {
-      stringified = obj.$to_s();
-      if (typeof stringified !== 'string' && !stringified.$$is_string) {
-        stringified = Opal.fallback_to_s(obj);
-      }
-      return stringified;
-    }
-    else {
-      return obj.toString();
-    }
-  }
-
   Opal.last_promise = null;
   Opal.promise_unhandled_exception = false;
