+++ opal/runtime/runtime.rb
@@ -2,5 +2,6 @@ require 'runtime/boot'
 require 'runtime/const'
 require 'runtime/send'
 require 'runtime/hash'
+require 'runtime/string'
 require 'runtime/regexp'
 require 'runtime/helpers'
